it has some logical error.i dont know correct form general logic error list here: -i think keys enum have wrong set value Engine.cs line 2993)==m_Device.add_DeviceCreated(new EventHandler(Engine.OnDeviceCreated));
texture.cs line 1191)===//bool flag = ((((int)data.pvSrc) + (y * data.Pitch))[x << 1] & 0x8000) != 0;

macro.cs linr 84)===return (((this.m_Key == 0x11000) || (this.m_Key == 0x11001)) || (this.m_Key == 0x11002));

macro.cs line 15)=== public const Keys WheelDown = 0x11001;
macro.cs line 16)=== public const Keys WheelPress = 0x11002;
macro.cs line 17)=== public const Keys WheelUp = 0x11000;

GMacroEditorPanel.cs line 165,169,173

Renderer.cs line 1245,1262,1275)=== IntPtr ptr4 = (IntPtr) (texturedRef2 + 3);

Renderer.cs line 1439,1311)=== texturedRef2 = *((CustomVertex.TransformedColoredTextured**) (texturedRef2 + 1));

+++ Source - Disassembler/Client/Client/AosAttributes.cs
@@ -1,7 +1,5 @@
 ﻿namespace Client
 {
-    using System;
-
     public class AosAttributes
     {
         private ObjectPropertyList m_Props;
@@ -15,7 +13,7 @@
         {
             foreach (ObjectProperty property in this.m_Props.Properties)
             {
-                if (property.Number == attr)
+                if (property.Number == (int)attr)
                 {
                     string arguments = property.Arguments;
                     switch (arguments)
@@ -378,4 +376,3 @@
         }
     }
 }
-
