+++ Source - Disassembler/Client/Client/ItemFlags.cs
@@ -1,8 +1,5 @@
 ﻿namespace Client
 {
-    using System;
-    using System.Reflection;
-
     public class ItemFlags
     {
         private int m_Value;
@@ -13,25 +10,25 @@
             {
                 return string.Format("Unknown flags: 0x{0:X2}", this.m_Value);
             }
-            return ((ItemFlag) this.m_Value).ToString();
+            return ((ItemFlag)this.m_Value).ToString();
         }
 
         public bool this[ItemFlag flag]
         {
             get
             {
-                int num = (int) flag;
-                return ((this.m_Value & flag) != 0);
+                int num = (int)flag;
+                return ((this.m_Value & (int)flag) != 0);
             }
             set
             {
                 if (value)
                 {
-                    this.m_Value |= flag;
+                    this.m_Value |= (int)flag;
                 }
                 else
                 {
-                    this.m_Value &= ~flag;
+                    this.m_Value &= (int)~flag;
                 }
             }
         }
@@ -55,4 +52,3 @@
         }
     }
 }
-
