+++ Source - Disassembler/Client/Client/TileFlags.cs
@@ -1,13 +1,12 @@
 ﻿namespace Client
 {
-    using System;
-    using System.Reflection;
     using System.Runtime.InteropServices;
 
     [StructLayout(LayoutKind.Sequential)]
     public struct TileFlags
     {
         public int Value;
+
         public TileFlags(int Value)
         {
             this.Value = Value;
@@ -15,16 +14,15 @@
 
         public override string ToString()
         {
-            return ((TileFlag) this.Value).ToString();
+            return ((TileFlag)this.Value).ToString();
         }
 
         public bool this[TileFlag flag]
         {
             get
             {
-                return ((this.Value & flag) != 0);
+                return ((this.Value & (int)flag) != 0);
             }
         }
     }
 }
-
